implement netowrk & protocol error handling
This makes the IRC connection and disconnection logic much more
reliable. Using the context manager makes it so CircuitPython will
automatically close the socket properly in case of a keyboard
interrupt or an autoreload. Previously, I was getting a lot of 433
Nick already in use errors due to not closing sockets properly.