add preformatted text bot command
You can send a line like this:

    !pre /line 1/line 2/line 3/line 4

and it will get rendered to the TFT display like this:

    line 1
    line 2
    line 3
    line 4

The first character after "!pre " is used as a delimeter that gets
replaced with a newline character. This compensates for IRC not
having a standard way to encode line breaks in messages.