switch from PRIVMSG to 332/TOPIC
This makes it so the bot uses the channel topic as its message
source rather than watching for PRIVMSG. TOPIC is sticky, so this
makes it so the bot can get the latest sensor summary report as
soon as it JOINs the IRC sensors channel. The old PRIVMSG way made
you wait multiple minutes for an initial update after a reset.