fix faulty reconnect logic
The code to reconnect in the event of nick in use error was busted
and could get stuck in an infinite loop. Also, the wifi reconnect
logic could attempt to reconnect when it was already connected.

This commit fixes those bugs and gets the exponential backoff retry
interval working properly for nick in use errors.